+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/Text.kt
@@ -19,6 +19,8 @@ public fun Text(
 	color: Color = Color.Unspecified,
 	background: Color = Color.Unspecified,
 	textStyle: TextStyle = TextStyle.Unspecified,
+	underlineStyle: UnderlineStyle = UnderlineStyle.Unspecified,
+	underlineColor: Color = Color.Unspecified,
 ) {
 	val layout = remember { StringTextLayout() }
 	layout.value = value
@@ -33,7 +35,7 @@ public fun Text(
 		},
 		modifier = modifier.drawBehind {
 			layout.lines.forEachIndexed { row, line ->
-				drawText(row, 0, line, color, background, textStyle)
+				drawText(row, 0, line, color, background, textStyle, underlineStyle, underlineColor)
 			}
 		},
 	)
@@ -47,6 +49,8 @@ public fun Text(
 	color: Color = Color.Unspecified,
 	background: Color = Color.Unspecified,
 	textStyle: TextStyle = TextStyle.Unspecified,
+	underlineStyle: UnderlineStyle = UnderlineStyle.Unspecified,
+	underlineColor: Color = Color.Unspecified,
 ) {
 	val layout = remember { AnnotatedStringTextLayout() }
 	layout.value = value
@@ -61,7 +65,7 @@ public fun Text(
 		},
 		modifier = modifier.drawBehind {
 			layout.lines.forEachIndexed { row, line ->
-				drawText(row, 0, line, color, background, textStyle)
+				drawText(row, 0, line, color, background, textStyle, underlineStyle, underlineColor)
 			}
 		},
 	)